Fix Player.getRatio() rounding issue
Player.getRatio() was returning a rounded result due to the division of
both its values being integers. Both integers are now cast to floats
before being divided; thus fixing the issue.

+++ src/sam/ferg/fantasyfootball/Player.java
@@ -41,8 +41,8 @@ public class Player {
 		return numberOfPlayoffs;
 	}
 	
-	public double getRatio() {
-		return wins/losses;
+	public float getRatio() {
+		return ((float) wins/ (float) losses);
 	}
 	
 	public void addWin() {
